Add integration tests for dynamic AAVE API
