+++ examples/dynamic-aave-agent/package.json
@@ -13,6 +13,7 @@
     "ethers": "^5.7.2",
     "openai": "^4.20.0",
     "readline": "^1.3.0",
+    "ts-pattern": "^5.7.0",
     "tsx": "^4.6.2"
   },
   "devDependencies": {
